QR-482: GitHub CVE Scanning
Update build script to check if subdirectory exists.

+++ build-codeql.ps1
@@ -1,6 +1,8 @@
 #! /usr/bin/pwsh
 
-Remove-Item -Path .\build -Recurse -Force
+if (Test-Path -Path .\build) {
+    Remove-Item -Path .\build -Recurse -Force
+}
 
 New-Item -Path .\build -ItemType Directory
